\name{Met.FID}
\alias{Met.FID}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Interactive Phase Correction and Peak reference}
\description{
  Interactive graphical display that allow zero order and first order phase corrections.
In addition the user can set the chemical shift of a certain compound (TSP or DSS) to 0 ppm as peak reference.
}
\usage{
Met.FID()
}
\details{
  The Met.FID graphical Display runs into the NMRS main function.
}
\value{
  \item{spect.fid }{Preprocessed spectra}
  
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }

\seealso{\link{Fid}, \link{phase}, \link{NMRS}}

\keyword{ interface }