\name{Met.spectrum}
\alias{Met.spectrum}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Spectrum Graphical Display }
\description{
  Interactive Graphical Display
}
\usage{
Met.spectrum(xCoords)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xCoords}{ Data frame to plot}
}
\details{
  Internal Function
}
\author{  Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} }
\seealso{\link{Met.FID}, \link{NMRS}}
 
\keyword{ interface }
