\name{NMRS-package}
\alias{NMRS-package}
\docType{package}
\title{
NMR Spectra Preprocessing
}
\description{
 This ''NMRs'' package permits the pre-processing of the raw NMR Bruker spectra. Data pre-processing transforms the data in a way that subsequent analysis and modelling are easier, more robust and more accurate.
 For the analysis of NMR spectra, pre-processing methods usually intend to reduce variances and influences as phase corrections of each spectrum, baseline corrections, etc. }
\details{
\tabular{ll}{
Package: \tab NMRS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-02-11\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
The NMRS  package has been designed as an interactive process. By Typing NMRS () the user has access to complete  preprocessing of  the data
}
\author{
Jose Luis Izquierdo (Instituto de Estudios Biofuncionales UCM)

Maintainer: Jose L. Izquierdo \email{izquierdo@ieb.ucm.es} 
}

\keyword{ package }
\keyword{ interface }

